+++ test/mock.js
@@ -1,7 +1,0 @@
-
-export const location = {
-  some: 'stuff',
-  more: 'things',
-  hash: '#xk',
-  pathname: '/foo',
-}